Remove partition test as its a lie with the mock driver
Kafka now admits that with th etopologytest driver, partitioning doesn't
really happen - so they have removed the partition field from the test
records used in the test driver. As a result we can't (and shouldn't)
test partitioning with the test driver - the test was always self
fulfilling anyway! Horrible as it feels - deleting the test.